[Impl] 술BTI Activity
