--- app/src/main/java/com/kud/hanzan/ui/sbti/SbtiProgressBarFragment.kt
@@ -1,9 +1,0 @@
-package com.kud.hanzan.ui.sbti
-
-import com.kud.hanzan.R
-import com.kud.hanzan.databinding.FragmentSbtiProgressBarBinding
-import com.kud.hanzan.utils.base.BaseFragment
-
-class SbtiProgressBarFragment() : BaseFragment<FragmentSbtiProgressBarBinding>(R.layout.fragment_sbti_progress_bar){
-
-}
