+++ app/src/main/java/com/kud/hanzan/ui/sbti/SbtiResultActivity.kt
@@ -1,0 +1,10 @@
+package com.kud.hanzan.ui.sbti
+
+import com.kud.hanzan.R
+import com.kud.hanzan.databinding.ActivitySbtiResultBinding
+import com.kud.hanzan.utils.base.BaseActivity
+
+class SbtiResultActivity : BaseActivity<ActivitySbtiResultBinding>(R.layout.activity_sbti_result) {
+    override fun initView() {
+    }
+}
