--- app/src/main/res/layout/fragment_sbti_progress_bar.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-        <variable
-            name="progress"
-            type="java.lang.Integer" />
-    </data>
-
-    <androidx.appcompat.widget.LinearLayoutCompat
-        tools:context=".ui.sbti.SbtiProgressBarFragment"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:background="@drawable/bg_sbti_progress_bar"
-        android:orientation="horizontal"
-        android:gravity="center">
-
-        <ProgressBar
-            style="@android:style/Widget.ProgressBar.Horizontal"
-            android:layout_width="250dp"
-            android:layout_height="6dp"
-            android:max="100"
-            android:min="0"
-            android:progress="20"
-            android:progressBackgroundTint="@color/text_primary"
-            android:progressTint="@color/color_main"/>
-
-        <androidx.appcompat.widget.AppCompatTextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:text="20%"
-            android:textSize="16sp"
-            android:textStyle="bold"/>
-    </androidx.appcompat.widget.LinearLayoutCompat>
-</layout>
