+++ app/src/main/res/values/strings.xml
@@ -56,6 +56,7 @@
     <string name="sbti_question_11">술의 맛은 어떤가요?</string>
     <string name="sbti_answer_11_1">달다.</string>
     <string name="sbti_answer_11_5">쓰다.</string>
+    <string name="navigate_to_sbti_result">결과 확인하기!</string>
     <!--    Main Activity-->
     <string name="app_name">한잔</string>
     <string name="home_search_hint">지금 당장 먹고 싶은 술을 검색!</string>
